clean up update
