+++ services/app-api/handlers/reports/archive.test.ts
@@ -1,0 +1,80 @@
+import { fetchReport } from "./fetch";
+import { updateReport } from "./update";
+import { APIGatewayProxyEvent } from "aws-lambda";
+import { proxyEvent } from "../../utils/testing/proxyEvent";
+import { StatusCodes } from "../../utils/types/types";
+import { mockReport } from "../../utils/testing/setupJest";
+import error from "../../utils/constants/constants";
+
+jest.mock("../../utils/auth/authorization", () => ({
+  isAuthorized: jest.fn().mockResolvedValue(true),
+  hasPermissions: jest.fn(() => {}),
+}));
+const mockAuthUtil = require("../../utils/auth/authorization");
+
+jest.mock("../../utils/debugging/debug-lib", () => ({
+  init: jest.fn(),
+  flush: jest.fn(),
+}));
+
+jest.mock("./fetch");
+const mockedFetchReport = fetchReport as jest.MockedFunction<
+  typeof fetchReport
+>;
+
+const mockProxyEvent: APIGatewayProxyEvent = {
+  ...proxyEvent,
+  headers: { "cognito-identity-id": "test" },
+  pathParameters: { state: "AB", id: "testReportId" },
+  body: JSON.stringify(mockReport),
+};
+
+const archiveEvent: APIGatewayProxyEvent = {
+  ...mockProxyEvent,
+  body: JSON.stringify({
+    ...mockReport,
+    archived: true,
+  }),
+};
+
+describe("Test archiveReport method", () => {
+  beforeEach(() => {
+    // fail state and pass admin auth checks
+    mockAuthUtil.hasPermissions
+      .mockReturnValueOnce(false)
+      .mockReturnValueOnce(true);
+  });
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  test("Test archive report passes with valid data", async () => {
+    mockedFetchReport.mockResolvedValue({
+      statusCode: 200,
+      headers: {
+        "Access-Control-Allow-Origin": "string",
+        "Access-Control-Allow-Credentials": true,
+      },
+      body: JSON.stringify(mockReport),
+    });
+    const res: any = await updateReport(archiveEvent, null);
+
+    const body = JSON.parse(res.body);
+    expect(res.statusCode).toBe(StatusCodes.SUCCESS);
+    expect(body.archived).toBe(true);
+  });
+
+  test("Test archive report with no existing record throws 404", async () => {
+    mockedFetchReport.mockResolvedValue({
+      statusCode: 200,
+      headers: {
+        "Access-Control-Allow-Origin": "string",
+        "Access-Control-Allow-Credentials": true,
+      },
+      body: undefined!,
+    });
+    const res = await updateReport(archiveEvent, null);
+    expect(res.statusCode).toBe(StatusCodes.NOT_FOUND);
+    expect(res.body).toContain(error.NO_MATCHING_RECORD);
+  });
+});
