+++ services/app-api/handlers/reports/update.test.ts
@@ -40,14 +40,6 @@ const updateEvent: APIGatewayProxyEvent = {
   }),
 };
 
-const archiveEvent: APIGatewayProxyEvent = {
-  ...mockProxyEvent,
-  body: JSON.stringify({
-    ...mockReport,
-    archived: true,
-  }),
-};
-
 const submissionEvent: APIGatewayProxyEvent = {
   ...mockProxyEvent,
   body: JSON.stringify({
@@ -178,47 +170,5 @@ describe("Test updateReport API method", () => {
 
     expect(res.statusCode).toBe(500);
     expect(res.body).toContain(error.NO_KEY);
-  });
-});
-
-describe("Test archiveReport method", () => {
-  beforeEach(() => {
-    // fail state and pass admin auth checks
-    mockAuthUtil.hasPermissions
-      .mockReturnValueOnce(false)
-      .mockReturnValueOnce(true);
-  });
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-
-  test("Test archive report passes with valid data", async () => {
-    mockedFetchReport.mockResolvedValue({
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin": "string",
-        "Access-Control-Allow-Credentials": true,
-      },
-      body: JSON.stringify(mockReport),
-    });
-    const res: any = await updateReport(archiveEvent, null);
-
-    const body = JSON.parse(res.body);
-    expect(res.statusCode).toBe(StatusCodes.SUCCESS);
-    expect(body.archived).toBe(true);
-  });
-
-  test("Test archive report with no existing record throws 404", async () => {
-    mockedFetchReport.mockResolvedValue({
-      statusCode: 200,
-      headers: {
-        "Access-Control-Allow-Origin": "string",
-        "Access-Control-Allow-Credentials": true,
-      },
-      body: undefined!,
-    });
-    const res = await updateReport(archiveEvent, null);
-    expect(res.statusCode).toBe(StatusCodes.NOT_FOUND);
-    expect(res.body).toContain(error.NO_MATCHING_RECORD);
   });
 });
